feat: Use lodash for object transformation
