+++ package.json
@@ -19,6 +19,7 @@
     "@types/node": "^12.7.11",
     "dts-bundle-webpack": "^1.0.2",
     "jest": "^24.8.0",
+    "lodash": "^4.17.15",
     "ts-jest": "^24.1.0",
     "ts-loader": "^6.2.0",
     "tslint": "^5.20.0",
